Created struct RevisionVertex. Added enum VertexColor. Removed enums to revisionvertex.h.
